% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CSCMatrix-class}
\alias{CSCMatrix-class}
\title{CSCMatrix}
\description{
The structure of \code{CSCMatrix} is the same
as the structure of \code{dgCMatrix}. However, the
\code{CSCMatrix} has weaker constraints compared to \code{dgCMatrix}.

\code{CSCMatrix} supportes limited operators. The user should convert it to
\code{dgCMatrix} if more comprehensive computation is required.
}
\details{
The \code{CSCMatrix} validates two constraints used in \code{dgCMatrix}:
\itemize{
 \item The row indices should be sorted with columns.
 \item The row indices should be unique with columns.
}
Therefore, the result of matrix-vector should be the same.
}
\examples{
# construct a CSCMatrix
m <- hashed.model.matrix(~ ., CO2, 8)
# convert it to dgCMatrix
m2 <- as(m, "dgCMatrix")
}
\seealso{
\code{\link{dgCMatrix-class}}
}

