% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback.R
\name{generate_split_callback}
\alias{generate_split_callback}
\title{Generate callback of split}
\usage{
generate_split_callback(input, delim = ",", type = c("existence", "count"))
}
\arguments{
\item{input}{character vector. The input of split}

\item{delim}{string. \code{delim} will be used as delimiter for splitting}

\item{type}{string. One of \code{c("existence", "count")}
"count" indicates the number of occurrence of the token. "existence" indicates the boolean that whether the token exist or not.}
}
\description{
Generate callback of split
}

