% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hashed.model.matrix}
\alias{hashed.model.matrix}
\title{Feature Hashing of Model Matrix}
\usage{
hashed.model.matrix(object, data = environment(object),
  contrasts.arg = NULL, xlev = NULL, transpose = TRUE,
  drop.unused.levels = FALSE, row.names = TRUE, verbose = FALSE,
  hash_size = 2^24, ...)
}
\arguments{
\item{hash_size}{integer value. The size of hashed variables.}

\item{object}{an object of an appropriate class.  For the default
    method, a model formula or terms object.}

\item{data}{a data frame created with \code{\link{model.frame}}.  If
    another sort of object, \code{model.frame} is called first.}

\item{contrasts.arg}{\describe{
      \item{for \code{sparse.model.matrix()}:}{A list, whose entries are
	contrasts suitable for input to the \code{\link{contrasts}}
	replacement function and whose names are the names of columns
	of \code{data} containing \code{\link{factor}}s.}
      \item{for \code{fac2Sparse()}:}{character string or \code{NULL} or
	(coercable to) \code{"\linkS4class{sparseMatrix}"}, specifying the
	contrasts to be applied to the factor levels.}
  }}

\item{xlev}{to be used as argument of \code{\link{model.frame}} if
    \code{data} has no \code{"terms"} attribute.}

\item{transpose}{logical indicating if the \emph{transpose} should be
    returned; if the transposed is used anyway, setting \code{transpose = TRUE}
    is more efficient.}

\item{drop.unused.levels}{should factors have unused levels dropped?
    The default for \code{sparse.model.matrix} has been changed to
    \code{FALSE}, 2010-07, for compatibility with \R's standard (dense)
    \code{\link{model.matrix}()}.}

\item{row.names}{logical indicating if row names should be used.}

\item{verbose}{logical or integer indicating if (and how much)
    progress output should be printed.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Feature Hashing of Model Matrix
}
\seealso{
\code{\link{sparse.model.matrix}}
}

