% Generated by roxygen2 (4.0.2): do not edit by hand
\name{interpret.tag}
\alias{interpret.tag}
\title{Interpret tag in formula}
\usage{
interpret.tag(object, data)
}
\arguments{
\item{object}{formula}

\item{data}{data.frame}
}
\description{
Interpret tag in formula
}
\examples{
\dontrun{
 data <- data.frame(a = c("1,2,3", "2,3,3", "1,3", "3"), type = c("a", "b", "a", "a"), stringsAsFactors = FALSE)
 r <- interpret.tag(~ tag(a, split = ",") * type, data)
}
}

