% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intToRaw}
\alias{intToRaw}
\title{Convert the integer to raw vector with endian correction}
\usage{
intToRaw(src)
}
\arguments{
\item{src}{integer value.}
}
\value{
raw vector with length 4
}
\description{
Convert the integer to raw vector with endian correction
}

