% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback.R
\name{ls_special}
\alias{ls_special}
\title{List the Registered Specials}
\usage{
ls_special()
}
\value{
character vector. The specials which could be used in the
formula interface.
}
\description{
List the Registered Specials
}

