% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tag}
\alias{tag}
\title{Expand concatenated feature}
\usage{
tag(x, split = ",", type = c("count", "existence"))
}
\arguments{
\item{x}{character vector or factor. The source of tag features.}

\item{split}{character vector. The split symbol for tag features.}

\item{type}{character value. Either "count" or "existence". "count" indicates the number of occurrence of the tag. "existence" indicates the boolean that whether the tag exist or not.}
}
\value{
integer vector for \code{type = "count"} and logical vector for \code{type = "existence"}.
}
\description{
Expand concatenated feature
}

