% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{test.tag}
\alias{test.tag}
\title{test.tag}
\format{For each element, the string represents the occurrence
of different tags. For example, the string "1,27,19,25,tp,tw"
of the first instance represents that the feature `1` is TRUE, the feature `27` is
TRUE, et. al. On the contrary, the missing feature such as `2`
is FALSE.}
\usage{
test.tag
}
\description{
This is a vector to demo the concatenated feature.
}
\keyword{datasets}

