% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_callback}
\alias{test_callback}
\title{Test the callback function.}
\usage{
test_callback(Rcallback, input)
}
\arguments{
\item{Rcallback}{external pointer. The pointer of the callback function.}

\item{input}{string. The input.}
}
\value{
character
}
\description{
Test the callback function.
}
\details{
The Rcallback is an external pointer which points to a functional pointer..
The signature of the functional pointer should be:
\code{std::vector<std::string> (*f)(const char* str)}
}

