% Generated by roxygen2 (4.0.1): do not edit by hand
\name{hashed.model.matrix}
\alias{hashed.model.matrix}
\title{Create a Model Matrix with Feature Hashing}
\usage{
hashed.model.matrix(object, data = environment(object), hash_size = 2^24,
  keep.hashing_mapping = FALSE)
}
\arguments{
\item{object}{formula. A symbolic description of the model.
The details of model specification are given under 'Details'.}

\item{data}{data.frame. The original data.}

\item{hash_size}{positive integer. The hash size of feature hashing.}

\item{keep.hashing.mapping}{logical value.
The indicator of whether store the hash mapping.}
}
\description{
Create a Model Matrix with Feature Hashing
}

