% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ipinyou}
\alias{ipinyou}
\alias{ipinyou.train}
\alias{ipinyou.test}
\title{iPinYou Real-Time Bidding Dataset for Computational Advertising Research}
\format{
The column name of the data is the description of the data in Zhang, Yuan, Wang, et al. (2014). 
Most of the columns should be clearly described by their column names. 
For the details of the dataset, please read the Zhang, Yuan, Wang, et al. (2014).

\code{BidID}, the id of the RTB which is the unique identifier of the events.

\code{Adid}, the advertiser id.

\code{UserTag}, the user tags (segments) in iPinYou's proprietary audience database. 
This is also a real example of the concatenated feature.
}
\source{
\url{http://data.computational-advertising.org/}
}
\usage{
data(ipinyou)
}
\description{
This is a sample from the iPinYou Real-Time Bidding dataset.
The data.frame named \code{ipinyou.train} is a sample from the data of 2013-10-19 and 
the data.frame named \code{ipinyou.test} is a sample from the  data of 2013-10-20.
}
\references{
W. Zhang, S. Yuan, J. Wang, et al. 
"Real-Time Bidding Benchmarking with iPinYou Dataset". 
In: arXiv preprint arXiv:1407.7073 (2014).
}
