% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hashed.model.matrix.R
\name{register_callback}
\alias{register_callback}
\title{Register Special Function for Formula Interface}
\usage{
register_callback(special, callback)
}
\arguments{
\item{special}{string. The name which will be used in formula interface.}

\item{callback}{external pointer. The pointer to the callback function. Please see the details.}
}
\description{
Register Special Function for Formula Interface
}

