% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callback.R
\name{register_callback}
\alias{register_callback}
\title{Register Special Function for Formula Interface}
\usage{
register_callback(special, callback_generator)
}
\arguments{
\item{special}{string. The name which will be used in formula interface.}

\item{callback_generator}{function which will create a callback. Please see the details.}
}
\description{
Register Special Function for Formula Interface
}
\examples{
register_callback("split", generate_split_callback)
}

